+++ ktlint-reporter-checkstyle/src/test/kotlin/com/pinterest/ktlint/reporter/checkstyle/CheckStyleReporterTest.kt
@@ -7,7 +7,6 @@ import java.io.ByteArrayOutputStream
 import java.io.PrintStream
 
 class CheckStyleReporterTest {
-
     @Test
     fun testReportGeneration() {
         val out = ByteArrayOutputStream()