+++ ktlint-reporter-json/src/main/kotlin/com/pinterest/ktlint/reporter/json/JsonReporter.kt
@@ -7,7 +7,6 @@ import java.util.ArrayList
 import java.util.concurrent.ConcurrentHashMap
 
 public class JsonReporter(private val out: PrintStream) : Reporter {
-
     private val acc = ConcurrentHashMap<String, MutableList<LintError>>()
 
     override fun onLintError(