+++ ktlint-reporter-json/src/test/kotlin/com/pinterest/ktlint/reporter/json/JsonReporterTest.kt
@@ -7,7 +7,6 @@ import java.io.ByteArrayOutputStream
 import java.io.PrintStream
 
 class JsonReporterTest {
-
     @Test
     fun testReportGeneration() {
         val out = ByteArrayOutputStream()